Setup TypeScript (#10)
