+++ bot/package.json
@@ -4,15 +4,18 @@
   "description": "",
   "main": "index.js",
   "scripts": {
+    "type": "tsc",
+    "lint": "tsc; eslint --fix ./src/**/*.ts",
+    "build": "babel --extensions '.ts' build -d build",
     "test": "echo \"Error: no test specified\" && exit 1",
     "dev": "nodemon index.js",
-    "start": "node index.js",
-    "lint": "eslint --fix"
+    "start": "node build/index.js"
   },
   "keywords": [],
-  "author": "Ayushman Chhabra <ayushmxn@outlook.com> (https://ayushmxn.me)",
+  "author": "Ayushman Chhabra <ayushmxn@outlook.com> (https://ayushmanchhabra.com)",
   "license": "MIT",
   "dependencies": {
+    "@types/node": "^14.14.9",
     "discord.js": "^12.3.1",
     "dotenv": "^8.2.0",
     "express": "^4.17.1"
@@ -21,7 +24,19 @@
     "@typescript-eslint/eslint-plugin": "^4.7.0",
     "@typescript-eslint/parser": "^4.7.0",
     "eslint": "^7.13.0",
-    "nodemon": "^2.0.4"
+    "nodemon": "^2.0.4",
+    "typescript": "^4.1.2"
+  },
+  "babel": {
+    "presets": [
+      "@babel/preset-env",
+      "@babel/typescript"
+    ],
+    "plugins": [
+      [
+        "@babel/transform-runtime"
+      ]
+    ]
   },
   "eslintConfig": {
     "env": {
